The field blender file
This blender file contains the field for the competition, with all main parts of it separated by collections.  Download the file and delete what is not necessary, export to Unity in whatever format is needed.  Does not include the little objects that have to be picked up by the robot.