Added colliders
Added mesh colliders to the surface, walls, and boxes in the scene. These seem to only cover the faces of the shape, and each face had to be selected separately. So this is likely to be computationally expensive to render, as well as might be missing faces. Will mention this in meeting.