Skeleton code files
I wrote a skeleton program that we can use once more parts of the robot are prototyped. Namely a main program that will act as a driver and be in charge of running the other smaller programs. I also have a program with functions that will be used to handle lists of targets, and finding the next closest one and updating where to go to. This assumes it will be given a vector of targets that will be found using the cameras. I need to wait until the cameras are able to do this before I can fully finish it. I also need to wait until the driver board is finished so I can see what programming the motors will look like, and write code for moving the robot to its targets. there are lots of comments in the code detailing what the methods do.